Remove param checking from Lua APIs, consolidate modules
What remains of errors.lua folded into runtime.lua
GameObjAPI.lua combined with GameObj.lua
misc.lua, objects.lua, and types.lua folded into api.lua
