+++ Desktop/Code12/Code12/globals.lua
@@ -2,7 +2,7 @@
 --
 -- globals.lua
 --
--- The global state and utility functions for the Code 12 Lua Runtime.
+-- The global state and utility functions for the Code12 Lua Runtime.
 --
 -- (c)Copyright 2018 by David C. Parker
 -----------------------------------------------------------------------------------------
