+++ Desktop/Default Test/output.txt
@@ -15,5 +15,3 @@ Line 1
 Line 2
 Line 3
 Line 4
-Line 5
-Long key
